Improving readability of plot_hub
